LLM/Chatbot Evaluator/requirement.txt
streamlit
google-generativeai
python-dotenv
langchain
pypdf
pinecone-client
langchain_google_genai
pymupdf